Move ShortcutErrorType to the Editor
