+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/EditShortcutsWindow.cpp
@@ -4,18 +4,17 @@
 #include <common/Display/dpi_aware.h>
 #include <common/utils/EventLocker.h>
 #include <common/utils/winapi_error.h>
-
-#include <keyboardmanager/common/ShortcutErrorType.h>
 #include <keyboardmanager/common/MappingConfiguration.h>
 
-#include <KeyboardManagerState.h>
-#include <Dialog.h>
-#include <KeyDropDownControl.h>
-#include <LoadingAndSavingRemappingHelper.h>
-#include <ShortcutControl.h>
-#include <Styles.h>
-#include <UIHelpers.h>
-#include <XamlBridge.h>
+#include "KeyboardManagerState.h"
+#include "Dialog.h"
+#include "KeyDropDownControl.h"
+#include "LoadingAndSavingRemappingHelper.h"
+#include "ShortcutControl.h"
+#include "Styles.h"
+#include "UIHelpers.h"
+#include "XamlBridge.h"
+#include "ShortcutErrorType.h"
 
 using namespace winrt::Windows::Foundation;
