+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/KeyboardManagerEditorStrings.h
@@ -2,7 +2,7 @@
 
 #include "pch.h"
 
-#include <keyboardmanager/common/ShortcutErrorType.h>
+#include "ShortcutErrorType.h"
 
 namespace KeyboardManagerEditorStrings
 {
