+++ src/modules/keyboardmanager/KeyboardManagerEditorLibrary/ShortcutErrorType.h
