+++ src/modules/keyboardmanager/common/Helpers.cpp
@@ -5,7 +5,6 @@
 #include <common/interop/shared_constants.h>
 #include <common/utils/process_path.h>
 
-#include "ShortcutErrorType.h"
 #include "KeyboardManagerConstants.h"
 
 namespace Helpers
