+++ src/modules/keyboardmanager/common/KeyboardManagerCommon.vcxproj
@@ -54,7 +54,6 @@
     <ClCompile Include="Shortcut.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="ShortcutErrorType.h" />
     <ClInclude Include="Input.h" />
     <ClInclude Include="KeyboardEventHandlers.h" />
     <ClInclude Include="MappingConfiguration.h" />
